requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
diff-match-patch==20241021
Django==5.1.4
django-import-export==4.3.4
idna==3.10
pillow==11.1.0
requests==2.32.3
sqlparse==0.5.3
sslcommerz-lib==1.0
tablib==3.7.0
tzdata==2024.2
urllib3==2.3.0
